% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db_general.R
\name{connect_db_general}
\alias{connect_db_general}
\title{Create a DB connection to AdIntel (RPostgres)}
\usage{
connect_db_general(year, host = "10.147.18.200", user = "postgres", password)
}
\arguments{
\item{year}{a number.}

\item{host}{a string.}

\item{user}{a stringr.}

\item{password}{a string.}
}
\description{
Create a DB connection to AdIntel (RPostgres)
}
