% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db_year.R
\name{connect_db_year}
\alias{connect_db_year}
\title{Create a DB connection to AdIntel by year (RPostgres)}
\usage{
connect_db_year(year)
}
\arguments{
\item{year}{a number.}
}
\description{
Create a DB connection to AdIntel by year (RPostgres)
}
