% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_lst.R
\name{create_table_lst}
\alias{create_table_lst}
\title{Create list for CREATE TABLE}
\usage{
create_table_lst(
  table,
  columns = list(column_name = NULL, data_type = NULL, precision = NULL),
  unique = NULL,
  partition_by_range = NULL
)
}
\description{
Create list for CREATE TABLE
}
