% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_lst.R
\name{create_table_maker}
\alias{create_table_maker}
\title{Create list for CREATE TABLE}
\usage{
create_table_maker(
  table,
  column_name,
  data_type,
  precision = NULL,
  unique = NULL,
  .col_partitionn = NULL
)
}
\description{
Create list for CREATE TABLE
}
