% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcooper_table_format.R
\name{dbc_remove_schema_name}
\alias{dbc_remove_schema_name}
\title{Remove Schema Name in Front of Table Names}
\usage{
dbc_remove_schema_name(string, schema_name = "public")
}
\arguments{
\item{string}{string or character vector with table names.}

\item{schema_name}{string with name of schema.}
}
\description{
Remove Schema Name in Front of Table Names
}
\examples{
x <- "public.table1"
dbc_remove_schema_name(x)
}
