% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_utf8.R
\name{ensure_utf8}
\alias{ensure_utf8}
\title{Title}
\usage{
ensure_utf8(x, from = "latin1", to = "UTF-8")
}
\arguments{
\item{x}{a tibble}

\item{from}{encoding to convert from. Default is "latin1"}

\item{to}{encoding to convert to. Default is "UTF-8"}
}
\value{
a tibble where every character column is converted to "to" (e.g. "UTF-8")
}
\description{
Title
}
