% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adintel_tsv.R
\name{read_adintel_tsv}
\alias{read_adintel_tsv}
\title{Load Adintel Tsv file}
\usage{
read_adintel_tsv(file, col_names, col_types)
}
\arguments{
\item{file}{string.}

\item{col_names}{character vector.}

\item{col_types}{cols() definition}
}
\value{
a tibble.
}
\description{
Load Adintel Tsv file
}
