% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_drop_table.R
\name{send_drop_table}
\alias{send_drop_table}
\title{Send SQL DROP TABLE}
\usage{
send_drop_table(con, table_name)
}
\arguments{
\item{con}{a connection.}

\item{table_name}{a string.}
}
\description{
Send SQL DROP TABLE
}
