% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences_utilities.R
\name{split_occurrence_query}
\alias{split_occurrence_query}
\title{Utility functions to modify AdIntel Occurrences data}
\usage{
split_occurrence_query(con, new_table, old_table, media_type_id)
}
\arguments{
\item{con}{a connection.}

\item{new_table}{a string.}

\item{old_table}{a string.}

\item{media_type_id}{a number.}
}
\description{
Utility functions to modify AdIntel Occurrences data
}
