% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_charpad.R
\name{str_charpad}
\alias{str_charpad}
\title{Padding of strings}
\usage{
str_charpad(x, pad = " ", off_margin = 0)
}
\description{
Padding of strings
}
\examples{
stopifnot("pad must be string" = is.character(pad))

}
