% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_separate_AbCd.R
\name{str_separate_AbCd}
\alias{str_separate_AbCd}
\title{Transform String from AbcDe to abc_de}
\usage{
str_separate_AbCd(string, sep = "_", to_lower = TRUE)
}
\arguments{
\item{string}{a string or a character vector.}

\item{sep}{a string used as separator. Default is '_'.}

\item{to_lower}{logical. Default is TRUE and mutate 'string' to lower case.}
}
\description{
Transform String from AbcDe to abc_de
}
\examples{
x <- "AbcdEfgHi"
str_separate_AbCd(x)
}
