% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_separate_upper.R
\name{str_separate_upper}
\alias{str_separate_upper}
\title{Modify string from 'AaBbbCc' to 'aa_bbb_cc'}
\usage{
str_separate_upper(string, sep = "_", to_lower = TRUE)
}
\arguments{
\item{string}{a string.}

\item{sep}{a string.}

\item{to_lower}{logical.}
}
\value{
a string
}
\description{
Modify string from 'AaBbbCc' to 'aa_bbb_cc'
}
